Handle folders (#1)
# Motivation
Current approach only imports files to a folder `Imported (YYYY-MM-dd
HH:mm)`
While this might work well for cases where folders are not needed,
there's other cases where folders are required.

# Solution
In order to solve this, a root folder `Imported (YYYY-MM-dd HH:mm)` will
still be created, but all the folders from source will be imported to
destination. Then, each file is imported to the right folder in
destination as it is in source.